Update densities.utils function locality
* put functions that are in densities.utils in the module that uses
  them, if that is the only module using the function
* use __future__ annotations on files touched
* ignore useless warnings from nptyping
