+++ tox.ini
@@ -5,7 +5,7 @@ black_version = 22.0
 [testenv]
 extras = tests
 passenv = PYTHONPATH
-commands = pytest tests {posargs}
+commands = pytest -W "ignore::DeprecationWarning:nptyping.typing_" tests {posargs}
 
 [testenv:lint]
 passenv = PYTHONPATH
